requirements_dev.txt
asgiref==3.5.2
autopep8==2.0.0
dj-database-url==1.0.0
dj-static==0.0.6
Django==4.1.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.3.0
pycodestyle==2.9.1
PyJWT==2.6.0
python-decouple==3.6
pytz==2022.5
sqlparse==0.4.3
static3==0.7.0
tomli==2.0.1
